Add liveness probes where possible
IDM and NAT don't provide any debug service/port so use a simple TCP
check on the main port.
Notifications and Audit do currently not provide a debug port either. As
they are not listening on any port by default we don't currently implement
a useful liveness probe for them.