Make it possible to add labels per service.
This PR adds the possibility to add labels per service. This is useful
for affinity rules and selecting pods and deployments in a more flexible way.

To do this, I've added a new key `appSpecificConfig` to the scope viaworks
`ocis.appNames`. This can in the future also be used to simplify other
service specific templates.

Fixes #251